Committing html and css files
I have setup the ui mostly. There are a few things in the content that are not right. The text over the images might not show up correctly depending on how it comes in and I am not sure how to setup the columns correctly. Please take a look and let me know what you think.